Snap.Snaplet.Auth.Backends.JsonFile

Documentation

initJsonFileAuthManager

Arguments

:: AuthSettings
Authentication settings for your app

-> SnapletLens b SessionManager
Lens into a SessionManager auth snaplet will use

-> FilePath
Where to store user data as JSON

-> SnapletInit b (AuthManager b)

Initialize a JSON file backed AuthManager

mkJsonAuthMgr :: FilePath -> IO JsonFileAuthManager

Load/create a datafile into memory cache and return the manager.

This data type can be used by itself for batch/non-handler processing.