Snap.Snaplet.Config

Documentation

newtype AppConfig

AppConfig contains the config options for command line arguments in snaplet-based apps.

Constructors

AppConfig

Fields

appEnvironment :: Maybe String

appConfigTyCon :: TyCon

AppConfig has a manual instance of Typeable due to limitations in the tools available before GHC 7.4, and the need to make dynamic loading tractable. When support for earlier versions of GHC is dropped, the dynamic loader package can be updated so that manual Typeable instances are no longer needed.

appOpts :: AppConfig -> [OptDescr (Maybe (Config m AppConfig))]

Command line options for snaplet applications.

commandLineAppConfig :: MonadSnap m => Config m AppConfig -> IO (Config m AppConfig)

Calls snap-server's extendedCommandLineConfig to add snaplet options to the built-in server command line options.