Snap.Snaplet.Heist.Interpreted

Description

A module exporting only functions for using interpreted templates. If you import the main Snap.Snaplet.Heist module, it's easy to accidentally use the compiled render function even when you're using interpreted Heist. Importing only this module will make it harder to make mistakes like that.

Documentation

data Heist b

The state for the Heist snaplet. To use the Heist snaplet in your app include this in your application state and use heistInit to initialize it. The type parameter b will typically be the base state type for your application.

class HasHeist b where

A single snaplet should never need more than one instance of Heist as a subsnaplet. This type class allows you to make it easy for other snaplets to get the lens that identifies the heist snaplet. Here's an example of how the heist snaplet might be declared:

 data App = App { _heist :: Snaplet (Heist App) }
 makeLenses ''App

 instance HasHeist App where heistLens = subSnaplet heist

 appInit = makeSnaplet "app" "" Nothing $ do
     h <- nestSnaplet "heist" heist $ heistInit "templates"
     addConfig h heistConfigWithMyAppSplices
     return $ App h

Methods

heistLens :: SnapletLens (Snaplet b) (Heist b)

A lens to the Heist snaplet. The b parameter to Heist will typically be the base state of your application.

type SnapletHeist b m a = HeistT (Handler b b) m a

type SnapletISplice b = SnapletHeist b (Handler b b) Template

Initializer Functions

heistInit

Arguments

:: FilePath
Path to templates

-> SnapletInit b (Heist b)

The Initializer for Heist. This function is a convenience wrapper around heistInit' that uses defaultHeistState and sets up routes for all the templates. It sets up a "heistReload" route that reloads the heist templates when you request it from localhost.

heistInit'

Arguments

:: FilePath
Path to templates

-> HeistConfig (Handler b b)
Initial HeistConfig

-> SnapletInit b (Heist b)

A lower level Initializer for Heist. This initializer requires you to specify the initial HeistConfig. It also does not add any routes for templates, allowing you complete control over which templates get routed.

addTemplates

Arguments

:: HasHeist b

=> Snaplet (Heist b)

-> ByteString
The url prefix for the template routes

-> Initializer b v ()

Adds templates to the Heist HeistState. Other snaplets should use this function to add their own templates. The templates are automatically read from the templates directory in the current snaplet's filesystem root.

addTemplatesAt

Arguments

:: HasHeist b

=> Snaplet (Heist b)

-> ByteString
URL prefix for template routes

-> FilePath
Path to templates

-> Initializer b v ()

Adds templates to the Heist HeistState, and lets you specify where they are found in the filesystem. Note that the path to the template directory is an absolute path. This allows you more flexibility in where your templates are located, but means that you have to explicitly call getSnapletFilePath if you want your snaplet to use templates within its normal directory structure.

addConfig :: Snaplet (Heist b) -> HeistConfig (Handler b b) -> Initializer b v ()

Adds more HeistConfig data using mappend with whatever is currently there. This is the preferred method for adding all four kinds of splices as well as new templates.

getHeistState :: HasHeist b => Handler b v (HeistState (Handler b b))

More general function allowing arbitrary HeistState modification.

modifyHeistState

Arguments

:: HasHeist b

=> (HeistState (Handler b b) -> HeistState (Handler b b))
HeistState modifying function

-> Initializer b v ()

More general function allowing arbitrary HeistState modification.

withHeistState

Arguments

:: HasHeist b

=> (HeistState (Handler b b) -> a)
HeistState function to run

-> Handler b v a

Runs a function on with the Heist snaplet's HeistState.

Handler Functions

render

Arguments

:: HasHeist b

=> ByteString
Template name

-> Handler b v ()

Renders a template as text/html. If the given template is not found, this returns empty.

renderAs

Arguments

:: HasHeist b

=> ByteString
Content type to render with

-> ByteString
Template name

-> Handler b v ()

Renders a template as the given content type. If the given template is not found, this returns empty.

heistServe :: HasHeist b => Handler b v ()

A handler that serves all the templates (similar to serveDirectory). If the template specified in the request path is not found, it returns empty. Also, this function does not serve any templates beginning with an underscore. This gives you a way to prevent some templates from being served. For example, you might have a template that contains only the navbar of your pages, and you probably wouldn't want that template to be visible to the user as a standalone template. So if you put it in a file called "_nav.tpl", this function won't serve it.

heistServeSingle

Arguments

:: HasHeist b

=> ByteString
Template name

-> Handler b v ()

Handler for serving a single template (similar to fileServeSingle). If the given template is not found, this throws an error.

heistLocal

Arguments

:: HasHeist b

=> (HeistState (Handler b b) -> HeistState (Handler b b))
HeistState modifying function

-> Handler b v a
Handler to run

-> Handler b v a

Runs a handler with a modified HeistState. You might want to use this if you had a set of splices which were customised for a specific action. To do that you would do:

 heistLocal (bindSplices mySplices) handlerThatNeedsSplices

withSplices

Arguments

:: HasHeist b

=> Splices (SnapletISplice b)
Splices to bind

-> Handler b v a
Handler to run

-> Handler b v a

Runs an action with additional splices bound into the Heist HeistState.

renderWithSplices

Arguments

:: HasHeist b

=> ByteString
Template name

-> Splices (SnapletISplice b)
Splices to bind

-> Handler b v ()

Renders a template with a given set of splices. This is syntax sugar for a common combination of heistLocal, bindSplices, and render.

clearHeistCache :: Heist b -> IO ()

Clears data stored by the cache tag. The cache tag automatically reloads its data when the specified TTL expires, but sometimes you may want to trigger a manual reload. This function lets you do that.